sankey/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.11.0
google-auth==2.15.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
idna==3.4
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
six==1.16.0
urllib3==1.26.13
